+++ zenodot/src/dd/kms/zenodot/api/common/GeneralizedConstructor.java
@@ -1,0 +1,23 @@
+package dd.kms.zenodot.api.common;
+
+import javax.annotation.Nonnull;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+
+/**
+ * Wrapper around a {@link java.lang.reflect.Constructor}.
+ */
+public interface GeneralizedConstructor extends GeneralizedExecutable
+{
+	@Nonnull
+	@Override
+	default Constructor<?> getWrappedExecutable() {
+		return getWrappedConstructor();
+	}
+
+	@Nonnull
+	Constructor<?> getWrappedConstructor();
+
+	Object newInstance(Object ... initargs) throws InstantiationException, IllegalAccessException,
+		IllegalArgumentException, InvocationTargetException;
+}