+++ zenodot/src/dd/kms/zenodot/api/result/codecompletions/CodeCompletionMethod.java
@@ -1,10 +1,11 @@
 package dd.kms.zenodot.api.result.codecompletions;
 
+import dd.kms.zenodot.api.common.GeneralizedMethod;
 import dd.kms.zenodot.api.result.CodeCompletion;
 
 import java.lang.reflect.Method;
 
 public interface CodeCompletionMethod extends CodeCompletion
 {
-	Method getMethod();
+	GeneralizedMethod getMethod();
 }